tanta-web/requirements.txt
appdirs==1.4.3
argon2-cffi==16.3.0
bcrypt==3.1.3
celery==4.0.2
cffi==1.10.0
Django==1.11.1
django-activity-stream==0.6.3
django-celery-results==1.0.1
dj-database-url==0.4.2
djangorestframework==3.6.3
django-webpack-loader==0.5.0
gunicorn==19.7.0
mixpanel==4.3.2
mixpanel-celery==0.8.0
olefile==0.44
openexchangerates==0.1.1
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pycparser==2.17
pyparsing==2.2.0
redis==2.10.5
six==1.10.0
tox==2.7.0
whitenoise==3.3.0
